Improves data export for tables Adds a hidden column for the 'Name' field in several tables, specifically for export purposes.
This allows the 'Name' to be searchable in the table but also included in the exported file, while not being displayed in the table itself. This change is applied to the Eyepiece, Filter, Instrument, Lens, and Location tables.

Additionally, the 'Weather' column is hidden for export in the Location table.